refactor: 상단의 로그인/로그아웃 메뉴를 분리한다.
